ai_service/requirements.txt
fastapi==0.110.1
pydantic==2.12.5
uvicorn==0.25.0
google-generativeai==0.8.6
